create featurizer package
