+++ rasa/core/policies/form_policy.py
@@ -5,7 +5,7 @@ from rasa.constants import DOCS_URL_MIGRATION_GUIDE
 from rasa.core.actions.action import ACTION_LISTEN_NAME
 from rasa.core.domain import Domain, STATE
 from rasa.core.events import FormValidation
-from rasa.core.featurizers import TrackerFeaturizer
+from rasa.core.featurizers.tracker_featurizers import TrackerFeaturizer
 from rasa.core.interpreter import NaturalLanguageInterpreter, RegexInterpreter
 from rasa.core.policies.memoization import MemoizationPolicy
 from rasa.core.trackers import DialogueStateTracker
@@ -90,7 +90,7 @@ class FormPolicy(MemoizationPolicy):
         return lookup
 
     def recall(
-        self, states: List[STATE], tracker: DialogueStateTracker, domain: Domain,
+        self, states: List[STATE], tracker: DialogueStateTracker, domain: Domain
     ) -> Optional[Text]:
         # modify the states
         return self._recall_states(self._modified_states(states))
