+++ rasa/core/policies/rule_policy.py
@@ -4,12 +4,8 @@ from typing import List, Dict, Text, Optional, Any, Set, TYPE_CHECKING
 import json
 
 from rasa.core.events import FormValidation
-from rasa.core.domain import (
-    Domain,
-    InvalidDomain,
-    STATE,
-)
-from rasa.core.featurizers import TrackerFeaturizer
+from rasa.core.domain import Domain, InvalidDomain, STATE
+from rasa.core.featurizers.tracker_featurizers import TrackerFeaturizer
 from rasa.core.interpreter import NaturalLanguageInterpreter
 from rasa.core.policies.memoization import MemoizationPolicy
 from rasa.core.policies.policy import SupportedData
